Create test helper function
